refactor(tests): split types_misc.rs into types_xml.rs + types_lsn.rs
Separate XML and LSN tests into dedicated files for clearer organization.
