+++ tests/core/types_lsn.rs
@@ -1,26 +1,6 @@
 use bytes::BytesMut;
 use sentinel_driver::types::lsn::PgLsn;
-use sentinel_driver::types::xml::PgXml;
 use sentinel_driver::types::{FromSql, Oid, ToSql};
-
-// -- PgXml --
-
-#[test]
-fn test_xml_roundtrip() {
-    let val = PgXml("<root><item>hello</item></root>".to_string());
-    let mut buf = BytesMut::new();
-    val.to_sql(&mut buf).ok();
-    let decoded = PgXml::from_sql(&buf).ok();
-    assert_eq!(decoded, Some(val));
-}
-
-#[test]
-fn test_xml_oid() {
-    assert_eq!(PgXml(String::new()).oid(), Oid::XML);
-    assert_eq!(<PgXml as FromSql>::oid(), Oid::XML);
-}
-
-// -- PgLsn --
 
 #[test]
 fn test_lsn_roundtrip() {
@@ -66,10 +46,4 @@ fn test_lsn_decode_wrong_size() {
     // covers lsn.rs line 32: map_err for wrong byte count
     assert!(PgLsn::from_sql(&[0u8; 4]).is_err());
     assert!(PgLsn::from_sql(&[0u8; 10]).is_err());
-}
-
-#[test]
-fn test_xml_decode_empty() {
-    let decoded = PgXml::from_sql(&[]).ok();
-    assert_eq!(decoded, Some(PgXml(String::new())));
 }
